requirements.txt
matplotlib==3.10.3
numpy==2.1.3
pillow==11.2.1
streamlit==1.45.1
tensorflow==2.19.0